Create index.ts files to handle dependencies inside sub modules
