+++ traffic-sniffer/lib/request-metadata/intercepted-request/index.ts
@@ -1,0 +1,2 @@
+export { InterceptedRequest } from "./intercepted-request";
+export { RequestKey } from "./request-key";
