+++ traffic-sniffer/lib/request-metadata/intercepted-request/intercepted-request.ts
@@ -6,21 +6,7 @@ import {
   PathMetadataConfig,
   PathResponseData,
 } from "../../../types/types";
-
-export class RequestKey extends Object {
-  constructor(public readonly method: string, public readonly url: string) {
-    super();
-  }
-
-  static fromString(string: string) {
-    const [method, path] = string.split(" ");
-    return new this(method, path);
-  }
-
-  toString(): string {
-    return `${this.method} ${this.url}`;
-  }
-}
+import { RequestKey } from "./Request-key";
 
 export class InterceptedRequest {
   static readonly defaultConfig: PathMetadataConfig = {
