+++ traffic-sniffer/lib/request-metadata/intercepted-requests/index.ts
@@ -1,0 +1,1 @@
+export { InterceptedRequests } from "./intercepted-requests";
