+++ traffic-sniffer/lib/request-metadata/intercepted-requests/intercepted-requests.ts
@@ -1,7 +1,7 @@
 import { Request } from "express";
-import { InterceptedRequest, RequestKey } from "./intercepted-request";
+import { InterceptedRequest, RequestKey } from "../intercepted-request/";
 import { Invocation, PathResponseData } from "../../../types/types";
-import { RequestsMap } from "./Requests-map";
+import { RequestsMap } from "./requests-map";
 
 export class InterceptedRequests {
   private requests: RequestsMap;
