+++ traffic-sniffer/lib/request-metadata/intercepted-requests/requests-map.ts
@@ -1,4 +1,4 @@
-import { InterceptedRequest, RequestKey } from "./intercepted-request";
+import { InterceptedRequest, RequestKey } from "../intercepted-request";
 
 export class RequestsMap implements Map<RequestKey, InterceptedRequest> {
   size: number;
